app/requirements.txt
Flask==1.1.1
Jinja2==2.10.3
mysql-connector-python==8.0.19
